Revert "Refactor how some tasks are handled." Better refactoring of pretty much the same thing follows.
This reverts commit 2609df0622332a5c76344a25e9010389c09be698.
