+++ src/main/groovy/com/jvoegele/gradle/plugins/android/AndroidPlugin.groovy
@@ -5,10 +5,8 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.plugins.JavaPlugin
 
-import com.jvoegele.gradle.enhancements.JavadocEnhancement
-import com.jvoegele.gradle.tasks.android.AndroidSdkToolsFactory
-import com.jvoegele.gradle.tasks.android.AdbErrorException
-import com.jvoegele.gradle.tasks.android.AndroidPackageTask
+import com.jvoegele.gradle.enhancements.JavadocEnhancement 
+import com.jvoegele.gradle.tasks.android.AndroidPackageTask;
 import com.jvoegele.gradle.tasks.android.ProGuard
 import com.jvoegele.gradle.tasks.android.ProcessAndroidResources
 
@@ -33,7 +31,6 @@ class AndroidPlugin implements Plugin<Project> {
   private platformToolsDir // used since SDK r8, so check if it exists before using!
 
   private Project project
-  private AndroidSdkToolsFactory sdkTools
   private logger
 
   private androidProcessResourcesTask, proguardTask, androidPackageTask, 
@@ -46,7 +43,6 @@ class AndroidPlugin implements Plugin<Project> {
     project.plugins.apply(JavaPlugin.class)
 
     this.project = project
-    this.sdkTools = new AndroidSdkToolsFactory(project)
     this.logger = project.logger
 
     androidConvention = new AndroidPluginConvention(project)
@@ -176,7 +172,26 @@ class AndroidPlugin implements Plugin<Project> {
 
   /** closure should only contain calls to args; remember that the device arg has already been set! */
   private void adbExec(closure) {
-    sdkTools.adbexec.execute([:], closure)
+    def stdout = new ByteArrayOutputStream() // output is small, we can safely read it into memory
+    project.exec {
+      executable project.ant['adb']
+      if (project.ant['adb.device.arg']) {
+        args project.ant['adb.device.arg']
+      }
+
+      // the closure must run in context of this project.exec, so it must have the same delegate
+      closure.delegate = delegate
+      closure()
+
+      standardOutput = stdout
+    }
+
+    def reader = new InputStreamReader(new ByteArrayInputStream(stdout.toByteArray()))
+    reader.eachLine {
+      if (it.toLowerCase().contains("failure") || it.toLowerCase().contains("error")) {
+        throw new AdbErrorException(it.trim());
+      }
+    }
   }
 
   private void defineTaskDependencies() {
