+++ src/main/groovy/com/jvoegele/gradle/tasks/android/AndroidAntTask.groovy
@@ -14,5 +14,5 @@ abstract class AndroidAntTask {
     this.androidConvention = project.convention.plugins.android
   }
 
-  public abstract void execute(Map args = [:], Closure closure = null);
+  public abstract void execute(Map args);
 }
